+++ NEWS.md
@@ -53,6 +53,7 @@ New library features
 --------------------
 * Function composition now works also on one argument `∘(f) = f` (#34251)
 
+* `isapprox` (or `≈`) now has a one-argument "curried" method `isapprox(x)` which returns a function, like `isequal` (or `==`)` ([#32305]).
 * `Ref{NTuple{N,T}}` can be passed to `Ptr{T}`/`Ref{T}` `ccall` signatures ([#34199])